[TESTS] split core and plugins tests
