+++ tests/core/clients/test_Client.py
