+++ tests/core/parsers/frostbite/test_util.py
