+++ tests/core/parsers/frostbite2/test_util.py
