+++ tests/core/parsers/homefront/test_HomefrontParser.py
