+++ tests/core/parsers/test_arma2.py
