+++ tests/core/parsers/test_arma3.py
