+++ tests/core/parsers/test_bf3.py
