+++ tests/core/parsers/test_cod6.py
