+++ tests/core/parsers/test_csgo.py
