+++ tests/core/parsers/test_insurgency.py
