+++ tests/core/parsers/test_mohw.py
