+++ tests/core/storage/common.py
