+++ tests/core/test_b3.py
