+++ tests/core/test_config.py
