+++ tests/core/test_functions.py
