+++ tests/core/test_update.py
