--- tests/exclude_test.txt
@@ -1,10 +1,0 @@
-tests.plugins.test_callvote.CallvoteTestCase
-tests.plugins.test_callvote.Test_events
-tests.plugins.test_callvote.Test_commands
-tests.plugins.test_firstkill.FirstKillCase
-tests.plugins.test_firstkill.Test_events
-tests.plugins.test_jumper.JumperTestCase
-tests.plugins.test_jumper.Test_events
-tests.plugins.test_jumper.Test_commands
-tests.plugins.test_spawnkill.SpawnkillTestCase
-tests.plugins.test_spawnkill.Test_events
