--- tests/run.py
@@ -1,8 +1,0 @@
-"""
-Run the all the tests using nosetests.
-Eventual command line arguments will be passed over to nosetests.
-"""
-
-if __name__ == '__main__':
-    import nose
-    nose.main()
